PPP-7-1 Bug fixed (Link Home page buttons to Pages).
