+++ src/pages/HeaderPages/Home.js
@@ -1,4 +1,3 @@
-import { useHistory } from "react-router-dom";
 
 import IconWrapper from "../../components/HomePageComponents/Wrappers/IconWrapper";
 import ButtonWrapper from "../../components/HomePageComponents/Wrappers/ButtonWrapper";
@@ -8,24 +7,15 @@ import SitterIcon from "../../components/HomePageComponents/Buttons/SitterButton
 
 const Home = () => {
 
-  let hisotry = useHistory();
+  
 
-  const changeSitterRouteHandler = () => {
-    const newPath = "/Dogs";
-    hisotry.push(newPath);
-  }
-
-  const changeDogRouteHandler = () => {
-    const newPath = "/Sitters";
-    hisotry.push(newPath);
-  }
 
     return (
         <>
         <IconWrapper />
         <ButtonWrapper>
-          <SitterIcon onClick={changeSitterRouteHandler}/>
-          <DogIcon onClick={changeDogRouteHandler}/>
+          <SitterIcon />
+          <DogIcon />
         </ButtonWrapper>
         </>
       );
